+++ CHANGELOG.md
@@ -26,6 +26,7 @@ In an attempt to make this library more "idiomatic" some functions have been ren
  - Renamed `GeoJson` to `GeoJSON`.
  - Renamed `ToGeoJson` to `ToGeoJSON`.
  - Renamed `WriteChanges` to `ChangeResponse`, this is now a general type and can be used when dealing with changefeeds.
+ - Removed depth limit when encoding values using `Expr`
 
 ### Removed